# server/internal/docs/swagger.yaml
basePath: /
definitions:
  github_com_hybrid-app_server_internal_domainutil.DomainInput:
    properties:
      enabled:
        type: boolean
      position:
        type: integer
      url:
        type: string
    type: object
  github_com_hybrid-app_server_internal_httpx.Envelope:
    properties:
      code:
        description: 0 = 成功，非 0 = 业务错误码（沿用 HTTP 状态码语义）
        type: integer
      data:
        description: 业务数据
      message:
        description: 人类可读信息
        type: string
    type: object
  github_com_hybrid-app_server_internal_model.AdjustEvents:
    additionalProperties:
      type: string
    type: object
...
  github_com_hybrid-app_server_internal_service.GateResponse:
    properties:
      mode:
        description: '"A" | "B"'
        type: string
      openMode:
        description: 仅 mode=B 时有值："internal"（内开）| "external"（外开）
        type: string
      url:
        description: 仅 mode=B 时有值：B 面完整地址（主域名 + /?palcode=，客户端原样打开）
        type: string
    type: object
...
  internal_handler.createListingCampaignReq:
    properties:
      body:
        type: string
      deeplinkPath:
        type: string
      extraData:
        additionalProperties:
          type: string
        type: object
      imageUrl:
        type: string
      listingIds:
        items:
          type: integer
        type: array
      name:
        type: string
      title:
        type: string
    type: object
  internal_handler.createListingReq:
    properties:
      adjustAppToken:
        type: string
      adjustEvents:
        $ref: '#/definitions/github_com_hybrid-app_server_internal_model.AdjustEvents'
      afAppId:
        type: string
      afDevKey:
        type: string
      brandCode:
        type: string
      bundleId:
        type: string
      displayName:
        type: string
      name:
        type: string
      openMode:
        type: string
      palCode:
        type: string
      platform:
        type: string
      remark:
        type: string
      storeUrl:
        type: string
    type: object
...
  internal_handler.gateReq:
    properties:
      bundleId:
        type: string
      platform:
        type: string
      timezone:
        type: string
    type: object
...
  internal_handler.registerListingTokenReq:
    properties:
      bundleId:
        type: string
      deviceToken:
        type: string
      gateMode:
        description: 客户端上报的最近 AB 面判定结果 A/B（决定是否进 B 面推送受众）
        type: string
      model:
        type: string
      platform:
        type: string
    type: object
...
  internal_handler.sendCampaignReq:
    properties:
      dryRun:
        type: boolean
    type: object
...
  internal_handler.setListingDomainsReq:
    properties:
      domains:
        items:
          $ref: '#/definitions/github_com_hybrid-app_server_internal_domainutil.DomainInput'
        type: array
      inheritBrand:
        type: boolean
    type: object
  internal_handler.setListingGateReq:
    properties:
      countries:
        items:
          type: string
        type: array
      ipAllowCidrs:
        items:
          type: string
        type: array
      ipDenyCidrs:
        items:
          type: string
        type: array
      timezoneDeny:
        items:
          type: string
        type: array
    type: object
  internal_handler.testListingGateReq:
    properties:
      ip:
        type: string
      timezone:
        type: string
    type: object
...
  internal_handler.updateListingReq:
    properties:
      adjustAppToken:
        type: string
      adjustEvents:
        $ref: '#/definitions/github_com_hybrid-app_server_internal_model.AdjustEvents'
      afAppId:
        type: string
      afDevKey:
        type: string
      brandCode:
        type: string
      displayName:
        type: string
      gateEnabled:
        type: boolean
      name:
        type: string
      openMode:
        type: string
      palCode:
        type: string
      remark:
        type: string
      status:
        type: string
      storeUrl:
        type: string
    type: object
...
info:
  contact: {}
  description: hybrid-app 多渠道打包后台：渠道 CRUD / 图标管线 / 域名配置 / 运行时下发。
  title: 渠道中台 API
  version: "1.0"
paths:
...
  /api/app/listing/gate:
    post:
      consumes:
      - application/json
      description: |-
        服务端按请求真实 IP 查国家 + 校验时区/IP 规则，返回 mode=A（展示应用内容）或 mode=B（放行到配置 URL）。
        mode=B 时额外带 openMode（internal=内开/external=外开），告诉客户端用哪种方式打开该 URL；mode=A 时不含该字段。
        出于安全：不返回任何判定原因/国家/命中规则，审核方也无法从响应反推规则。
      parameters:
      - description: platform/bundleId/timezone
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.gateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_service.GateResponse'
      summary: AB 面判定（公开，客户端 App 启动调用，不缓存）
      tags:
      - app
  /api/app/listing/register-token:
    post:
      consumes:
      - application/json
      description: 客户端拿到 gate 判定后注册 FCM/APNs token 并上报 gateMode；上架包推送强制只发 B 面设备。
      parameters:
      - description: platform/bundleId/deviceToken/gateMode
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.registerListingTokenReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      summary: 上架包设备 push token 注册（公开，带 AB 面判定结果）
      tags:
      - app
...
  /api/listings:
    get:
      parameters:
      - description: android / ios
        in: query
        name: platform
        type: string
      - description: enabled/disabled/archived
        in: query
        name: status
        type: string
      - description: 关键词（名称/包名）
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 上架包列表（按平台/状态/关键词筛选，含品牌/域名/网关）
      tags:
      - listings
    post:
      consumes:
      - application/json
      parameters:
      - description: 上架包字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createListingReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 新增上架包（默认继承品牌域名、gate_enabled=false 只有 A 面）
      tags:
      - listings
  /api/listings/{id}:
    delete:
      parameters:
      - description: 上架包 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 删除上架包（连带域名/网关/日志级联删除）
      tags:
      - listings
    get:
      parameters:
      - description: 上架包 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 上架包详情
      tags:
      - listings
    put:
      consumes:
      - application/json
      parameters:
      - description: 上架包 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 可选更新字段
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.updateListingReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 修改上架包（打开 gateEnabled 前服务端强制校验国家白名单非空）
      tags:
      - listings
  /api/listings/{id}/domains:
    put:
      consumes:
      - application/json
      parameters:
      - description: 上架包 ID
        in: path
        name: id
        required: true
        type: integer
      - description: inheritBrand=true 表示切回继承
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.setListingDomainsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 设置上架包 B 面域名覆盖 / 切回继承品牌
      tags:
      - listings
  /api/listings/{id}/gate:
    put:
      consumes:
      - application/json
      parameters:
      - description: 上架包 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 网关规则
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.setListingGateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 保存上架包 AB 面网关规则（国家白名单必填、拒绝 CN/US）
      tags:
      - listings
  /api/listings/{id}/gate/logs:
    get:
      parameters:
      - description: 上架包 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 条数，默认 100，上限 500
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 上架包网关判定流水（排查「为什么没进 B 面」）
      tags:
      - listings
  /api/listings/{id}/gate/test:
    post:
      consumes:
      - application/json
      parameters:
      - description: 上架包 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 试算输入
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.testListingGateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 用指定 IP/时区试算 AB 面判定（后台自查，返回原因）
      tags:
      - listings
...
  /api/push/listing-campaigns:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 上架包推送活动列表（kind=listing）
      tags:
      - listings
    post:
      consumes:
      - application/json
      parameters:
      - description: 活动字段 + 目标上架包 ids
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.createListingCampaignReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 创建上架包推送活动（草稿；发送时强制只发 B 面设备）
      tags:
      - listings
  /api/push/listing-campaigns/{id}/send:
    post:
      consumes:
      - application/json
      parameters:
      - description: 活动 ID
        in: path
        name: id
        required: true
        type: integer
      - description: dryRun
        in: body
        name: body
        schema:
          $ref: '#/definitions/internal_handler.sendCampaignReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_hybrid-app_server_internal_httpx.Envelope'
      security:
      - BearerAuth: []
      summary: 发送上架包推送活动（dryRun 预览受众；真发只投 B 面设备）
      tags:
      - listings
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"